# api/openapi.yaml
openapi: 3.1.0
info:
  title: 127ohoh1 control-plane API (development reference)
  version: 0.1.0
  description: >
    Public reference API. Paths are added as phases land. Payment endpoints are
    simulated, development-only, and fail closed unless explicitly enabled.
    All timestamps are UTC RFC 3339. Errors use the envelope in `Error`.
servers:
  - url: http://localhost:8080
security:
  - session: []
paths:
...
  /v1/endpoints/{endpoint_id}/policy:
    parameters:
      - { name: endpoint_id, in: path, required: true, schema: { type: string } }
    get:
      summary: Resolved enforcement policy (owner or edge service credential)
      description: Contains limits and lifecycle state only; no price, asset, wallet or plan fields.
      responses:
        "200":
          description: Policy
          content: { application/json: { schema: { $ref: '#/components/schemas/EndpointPolicy' } } }
        "404": { $ref: '#/components/responses/Error' }
...
  /v1/policies/changes:
    get:
      summary: "Edge service: long-poll the policy change feed"
      security: [ { edge: [] } ]
      parameters:
        - { name: since, in: query, required: true, schema: { type: integer, minimum: 0 } }
        - { name: wait, in: query, schema: { type: integer, minimum: 0, maximum: 30 } }
      responses:
        "200":
          description: Changes (reset=true means refresh everything)
  /v1/hostnames/{label}:
    get:
      summary: "Edge service: what one tenant label is (an active reserved name may be issued a certificate before its first tunnel)"
      security: [ { edge: [] } ]
      parameters:
        - { name: label, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Hostname status
          content:
            application/json:
              schema:
                type: object
                required: [endpoint_id, kind, state]
                properties:
                  endpoint_id: { type: string }
                  kind: { type: string, enum: [random, reserved] }
                  state: { type: string, enum: [active, suspended, expired, released, pending] }
        "401": { $ref: '#/components/responses/Error' }
        "404": { $ref: '#/components/responses/Error' }
  /v1/admin/endpoints/{endpoint_id}/suspend:
    post:
      summary: "Administrative hook: suspend an endpoint (audited; disabled unless an admin token is configured)"
      security: [ { admin: [] } ]
      parameters:
        - { name: endpoint_id, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: Suspended }, "404": { $ref: '#/components/responses/Error' } }
  /v1/admin/endpoints/{endpoint_id}/unsuspend:
    post:
      summary: "Administrative hook: restore a suspended endpoint"
      security: [ { admin: [] } ]
      parameters:
        - { name: endpoint_id, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: Restored } }
  /v1/admin/principals/{principal_id}/suspend:
    post:
      summary: "Administrative hook: suspend a principal (all endpoints resolve to suspended)"
      security: [ { admin: [] } ]
      parameters:
        - { name: principal_id, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: Suspended } }
  /v1/admin/emergency:
    post:
      summary: "Administrative hook: global safety override (all endpoints resolve to suspended)"
      security: [ { admin: [] } ]
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [enabled], properties: { enabled: { type: boolean }, reason: { type: string, maxLength: 256 } } }
      responses: { "200": { description: Applied } }
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
    edge:
      type: http
      scheme: bearer
      description: Edge service credential (constant-time compared)
    admin:
      type: http
      scheme: bearer
      description: Administrative token; the admin API is absent (404) when none is configured
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            details: { type: object }
...
    EndpointPolicy:
      type: object
      required: [endpoint_id, policy_version, state, limits, usage, hostname, valid_until]
      properties:
        endpoint_id: { type: string }
        policy_version: { type: integer, description: monotonically increasing per endpoint }
        state: { type: string, enum: [active, suspended, expired, released, pending] }
        limits:
          type: object
          properties:
            bandwidth_bytes_per_second: { type: integer }
            burst_bytes: { type: integer }
            monthly_transfer_bytes: { type: integer }
            concurrent_requests: { type: integer }
            max_request_header_bytes: { type: integer }
            max_request_body_bytes: { type: integer }
        usage:
          type: object
          properties:
            period_start: { type: string, format: date-time }
            period_end: { type: string, format: date-time }
            used_bytes: { type: integer }
        hostname:
          type: object
          properties:
            kind: { type: string, enum: [random, reserved] }
            inactive_expiry_seconds: { type: [integer, "null"] }
        valid_until: { type: string, format: date-time }